# api/openapi.yaml
openapi: 3.1.0
info:
  title: Muxio API
  version: 0.1.0
  description: |
    Local HTTP API for the Muxio information capture core.

    The service listens on loopback only and has no authentication. Remote
    access requires a threat model first (ADR-002).
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Check process health
      description: Reports that the process is running. It does not check storage.
      responses:
        "200":
          description: The process is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /readyz:
    get:
      operationId: getReadiness
      summary: Check service readiness
      description: Reports whether the service can reach its database.
      responses:
        "200":
          description: The service is ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: The service cannot reach its database.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/status:
    get:
      operationId: getStatus
      summary: Read the local service status
      responses:
        "200":
          description: Current service status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /api/v1/sources:
    get:
      operationId: listSources
      summary: List configured sources
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Before"
      responses:
        "200":
          description: A page of sources, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcePage"
        "400":
          $ref: "#/components/responses/InvalidArgument"
  /api/v1/runs:
    get:
      operationId: listRuns
      summary: List collection runs
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Before"
        - name: source_id
          in: query
          description: Restrict the page to one source.
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: A page of runs, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunPage"
        "400":
          $ref: "#/components/responses/InvalidArgument"
  /api/v1/runs/{id}:
    get:
      operationId: getRun
      summary: Read one run
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{id}/events:
    get:
      operationId: listRunEvents
      summary: List the events explaining one run
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Before"
      responses:
        "200":
          description: A page of events, oldest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEventPage"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/config:
    get:
      operationId: getConfig
      summary: Read the effective configuration
      description: |
        Returns each setting's effective value together with where it came
        from. Credentials are stored outside the configuration file and never
        appear here (ADR-005).
      responses:
        "200":
          description: The effective configuration.
          headers:
            ETag:
              description: Fingerprint of the configuration file, for use with If-Match.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "500":
          $ref: "#/components/responses/Internal"
    put:
      operationId: replaceConfig
      summary: Replace the configuration file
      description: |
        Writes the whole configuration. Every known setting must be present:
        omitting one would silently reset it to its default.

        `If-Match` is required and must carry the `ETag` from a preceding GET.
        A stale value means someone else changed the file first, and the write
        is refused rather than overwriting their edit (ADR-005, ADR-006).

        The write lands in the file. A setting currently overridden by the
        environment keeps its overridden effective value, and the response
        reports its origin as `env`.
      parameters:
        - name: If-Match
          in: header
          required: true
          description: The ETag the edit was based on.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdate"
      responses:
        "200":
          description: The configuration was written.
          headers:
            ETag:
              description: Fingerprint of the newly written file.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "412":
          description: The file changed since the ETag was issued.
          headers:
            ETag:
              description: The current fingerprint.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The request body is too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match was not supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/Internal"
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Values above the maximum are clamped to it.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Before:
      name: before
      in: query
      description: Return only records with an id below this value.
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
    RunID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    InvalidArgument:
      description: A parameter was not acceptable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Internal:
      description: The request could not be completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Health:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
            - ready
            - unavailable
    Status:
      type: object
      additionalProperties: false
      required:
        - status
        - version
      properties:
        status:
          type: string
          const: ok
        version:
          type: string
    Error:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: A stable machine-readable code.
          enum:
            - invalid_argument
            - not_found
            - method_not_allowed
            - precondition_required
            - precondition_failed
            - payload_too_large
            - internal
        message:
          type: string
          description: A human-readable explanation.
        field:
          type: string
          description: The parameter or setting the problem applies to.
    Source:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - connector_kind
        - enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        connector_kind:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourcePage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_before
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        next_before:
          type: [integer, "null"]
          format: int64
          description: Cursor for the next page, or null when the page is the last.
    Run:
      type: object
      additionalProperties: false
      required:
        - id
        - source_id
        - source_name
        - trigger
        - status
        - started_at
        - finished_at
        - imported_count
        - duplicate_count
        - failed_count
        - attempt
        - last_error
      properties:
        id:
          type: integer
          format: int64
        source_id:
          type: integer
          format: int64
        source_name:
          type: string
        trigger:
          type: string
        status:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - partial
            - failed
            - canceled
            - interrupted
        started_at:
          type: string
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
        imported_count:
          type: integer
        duplicate_count:
          type: integer
        failed_count:
          type: integer
        attempt:
          type: integer
        last_error:
          type: [string, "null"]
    RunPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_before
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        next_before:
          type: [integer, "null"]
          format: int64
    RunEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - run_id
        - level
        - message
        - detail
        - occurred_at
      properties:
        id:
          type: integer
          format: int64
        run_id:
          type: integer
          format: int64
        level:
          type: string
          enum:
            - info
            - warn
            - error
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
        occurred_at:
          type: string
          format: date-time
    RunEventPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_before
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RunEvent"
        next_before:
          type: [integer, "null"]
          format: int64
    ConfigSetting:
      type: object
      additionalProperties: false
      required:
        - key
        - value
        - origin
      properties:
        key:
          type: string
          description: Dotted path such as server.addr.
        value:
          type: string
        origin:
          type: string
          description: Where the effective value came from.
          enum:
            - default
            - file
            - env
            - flag
    ConfigUpdate:
      type: object
      additionalProperties: false
      required:
        - settings
      properties:
        settings:
          type: array
          description: Every known setting. A missing one is an error, not a reset.
          items:
            type: object
            additionalProperties: false
            required:
              - key
              - value
            properties:
              key:
                type: string
              value:
                type: string
    Config:
      type: object
      additionalProperties: false
      required:
        - path
        - exists
        - settings
      properties:
        path:
          type: string
          description: The configuration file consulted, whether or not it exists.
        exists:
          type: boolean
        settings:
          type: array
          items:
            $ref: "#/components/schemas/ConfigSetting"